// src/models/office/LSS_T_MENU.model.ts
export interface LSS_T_MENU 
{
  CODE: string;
  LEVEL: number;
  WIDTH: number;
  SEQ: number;
  NAME: number;
  PARENT_CODE: string;
  DESCRIPTION: string;
  DISPLAY: string;
  ENABLE: string;
  URL_ACTION: string;
  URL_TYPE: string;
  URL_TARGET: string;
  CONDITION1: string;
  DEL: string;
  CREATE_USER: string;
  UPDATE_USER: string;
  ACCOUNT_TYPE_CODE: string;
  ADMIN_FLAG: string;
  DELETE: string;
  VIEW: string;
  SEARCH: string;
  EDIT: string;
  ADD: string;
  COMPANY_TYPE_CODE: string;
  CREATED_DATE: Date;
  UPDATED_DATE: Date;
}