// src/models/office/LSS_T_ACCOUNT.model.ts
export interface LSS_T_ACCOUNT {
  id?: any;
  CODE: string;
  ACCOUNT_TYPE_CODE: string;
  COMPANY_CODE: string;
  COMPANY_NAME:string;
  COMPANY_LICENSE: string;
  LOGIN_NAME: string;
  PREFIX_CODE: string;
  FIRSTNAME_EN: string;
  LASTNAME_EN: string;
  FIRSTNAME_TH: string;
  LASTNAME_TH: string;
  IDCARD: string;
  IDCARD_TYPE: string;
  COMPANY_POSITION: string;
  TEL: string;
  FAX: string;
  EMAIL: string;
  REMARK: string;
  RETRY_COUNT: string;
  BANNED: string;
  POSITION_CODE: string;
  SECTION_CODE: string;
  EMPLOYEE_CODE: string;
  GROUP_CODE: string;
  RECOVERY_TOKEN: string;
  GROUP_STAFF_CODE: string;
  IP: string;
  ENABLE: string;
  DEL: string;
  CREATE_USER: string;
  UPDATE_USER: string;
  STATUS: string;
  MOBILE: string;
  REASON: string;
  SIGNATOR1: string;
  SIGNATOR2: string;
  APPROVED_DATE: string;
  CREATED_DATE: string;
  LAST_PASSWORD_CHANGE_DATE: string;
  LAST_LOGIN_DATE: string;
  UPDATED_DATE: string;
  ACTIVED_DATE: string;
  EXPIRED_DATE: string;
  BANNED_DATE: string;
  START_DATE: string;
  END_DATE: string;
}