// src/models/office/LSS_V_LSSIRA10.model.ts
export interface LSS_V_LSSIRA10 {
  id: any;
  TRANS_DATE: Date;
  LOGIN_NAME: string;
  ACCOUNT_TYPE_NAME: string;
  ACCOUNT_TYPE_CODE: string;
  IP_ADDRESS: string;
  YSTEM_CODE: string;
  MENU_CODE: string;
  MENU_NAME: string;
  FUNCTION_NAME: string;
  LOG_HEADER: string;
  COMPANY_CODE: string;
  COMPANY_NAME: string;
  SECTION_CODE: string;
  SECTION_NAME: string;
  POSITION_CODE: string;
  POSITION_NAME: string;
  QTY: number;
}